[Feat] 성향테스트, 결과, 둘러보기 플로우 연결 (#128)
