--- ThunderRing/ThunderRing/Sources/Test/Cells/TempCell.swift
@@ -1,8 +1,0 @@
-//
-//  TempCell.swift
-//  ThunderRing
-//
-//  Created by soyeon on 2022/02/06.
-//
-
-import Foundation
