+++ ThunderRing/ThunderRing/Sources/Test/VCs/TestResultViewController.swift
@@ -21,7 +21,7 @@ final class TestResultViewController: UIViewController {
     }
     
     private lazy var contentView = UIView().then {
-        $0.backgroundColor = .white
+        $0.backgroundColor = .background
     }
     
     private var imageView = UIImageView().then {
@@ -44,7 +44,7 @@ final class TestResultViewController: UIViewController {
         $0.textColor = .gray100
         $0.font = .SpoqaHanSansNeo(type: .regular, size: 15)
         $0.numberOfLines = 0
-        $0.text = "자신만의 세계관을 만드는 것을 즐겨해서 상상력이 뛰어나죠. 아티스트의 피가 흘러 뚝딱뚝딱 만드는 모임을 좋아해요. 그래서인지 금방 그럴싸한 것들을 만들어 내곤하죠! 그래도 조심해주세요. 80데시벨 이상이 넘어가면 어지럽거든요."
+        $0.text = "궁금한 것들 투성이에요. 특히 사람들이 궁금해요. 그래서 모임에 나가면 사람들에게 먼저 말을 걸어요. 한마디로 분위기 메이커죠. 상황에 몰입을 잘해서 때론 과몰입이라는 소리를 듣곤 해요. 그래도 조심해 주세요. 제 몰입을 깨는 건 용납 못해요."
         $0.setTextSpacingBy(value: -0.6)
         $0.setLineSpacing(lineSpacing: 7, lineHeightMultiple: 0)
     }
@@ -101,7 +101,7 @@ final class TestResultViewController: UIViewController {
         $0.addTarget(self, action: #selector(touchUpSearchButton), for: .touchUpInside)
     }
     
-    private var retryButton = UIButton().then {
+    private var returnButton = UIButton().then {
         $0.setTitle("", for: .normal)
         $0.setImage(UIImage(named: "btn_return"), for: .normal)
         $0.initViewBorder(borderWidth: 1, borderColor: UIColor.purple100.cgColor, cornerRadius: 26, bounds: true)
@@ -140,7 +140,7 @@ final class TestResultViewController: UIViewController {
         view.addSubview(scrollView)
         scrollView.addSubview(contentView)
         
-        contentView.addSubviews([imageView, titleLabel, descriptionView, subTitleLabel, tendencyStackView, shareLabel, shareStackView, searchButton, retryButton])
+        contentView.addSubviews([imageView, titleLabel, descriptionView, subTitleLabel, tendencyStackView, shareLabel, shareStackView, searchButton, returnButton])
         descriptionView.addSubview(descriptionLabel)
         
         tendencyStackView.addArrangedSubview(firstTendencyView)
@@ -150,8 +150,7 @@ final class TestResultViewController: UIViewController {
         shareStackView.addArrangedSubview(facebookShareButton)
         
         scrollView.snp.makeConstraints {
-            $0.leading.trailing.top.equalToSuperview()
-            $0.bottom.equalTo(view.safeAreaLayoutGuide)
+            $0.leading.trailing.top.bottom.equalToSuperview()
         }
         
         contentView.snp.makeConstraints {
@@ -209,28 +208,23 @@ final class TestResultViewController: UIViewController {
         searchButton.snp.makeConstraints {
             $0.top.equalTo(shareStackView.snp.bottom).offset(55)
             $0.leading.equalToSuperview().inset(26)
-            $0.bottom.equalToSuperview().inset(17)
+            $0.bottom.equalToSuperview().inset(51)
             $0.width.equalTo(265)
             $0.height.equalTo(52)
         }
         
-        retryButton.snp.makeConstraints {
+        returnButton.snp.makeConstraints {
             $0.leading.equalTo(searchButton.snp.trailing).offset(7)
             $0.centerY.equalTo(searchButton.snp.centerY)
             $0.width.height.equalTo(52)
-            $0.bottom.equalToSuperview().inset(17)
+            $0.bottom.equalToSuperview().inset(51)
         }
-    }
-    
-    // MARK: - Custom Method
-    
-    private func updateLayout() {
-        
     }
     
     // MARK: - @objc
     
     @objc func touchUpSearchButton() {
+        NotificationCenter.default.post(name: NSNotification.Name("SearchPublicGroup"), object: nil)
         dismiss(animated: true)
     }
 }
