+++ ThunderRing/ThunderRing/Sources/Test/Views/LastTestView.swift
@@ -1,0 +1,17 @@
+//
+//  LastTestView.swift
+//  ThunderRing
+//
+//  Created by 소연 on 2022/05/16.
+//
+
+import UIKit
+
+import SnapKit
+import Then
+
+final class LastTestView: TestView {
+    
+    
+    
+}
